pyproject.toml
[project]
name = "llm-cookbook"
version = "1.0.0"
description = "LLM-CookBook 项目是一个面向开发者的大模型手册，针对国内开发者的实际需求，主打 LLM 全方位入门实践。本项目基于吴恩达老师大模型系列课程内容，对原课程内容进行筛选、翻译、复现和调优，覆盖从 Prompt Engineering 到 RAG 开发、模型微调的全部流程，用最适合国内学习者的方式，指导国内开发者如何学习、入门 LLM 相关项目。"
authors = [
    { name = "Datawhale", email = "opensource@datawhale.club" }
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "chromadb>=1.0.12",
    "datasets>=3.6.0",
    "docarray>=0.41.0",
    "google-search-results>=2.4.2",
    "gradio>=5.33.1",
    "huggingface-hub[hf-xet]>=0.32.4",
    "ipywidgets>=8.1.7",
    "jupyter-nbextensions-configurator>=0.6.4",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.21",
    "llama-index>=0.12.41",
    "matplotlib>=3.10.3",
    "notebook>=7.4.3",
    "numexpr>=2.10.2",
    "openai>=1.82.1",
    "openapi-pydantic>=0.5.1",
    "panel>=1.7.1",
    "pip>=25.1.1",
    "pyarrow>=20.0.0",
    "pypdf>=5.6.0",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "setuptools>=80.9.0",
    "torchvision>=0.22.1",
    "umap-learn>=0.5.7",
    "wandb>=0.20.1",
    "wheel>=0.45.1",
    "widgetsnbextension>=4.0.14",
    "wikipedia>=1.4.0",
]

[project.optional-dependencies]
full = [
]
test = []

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
